requirements.txt
asn1crypto==1.3.0
certifi==2019.11.28
cffi==1.13.2
cryptography==2.8
fpdf==1.7.2
idna==2.8
pdfkit==0.6.1
Pillow==7.0.0
pycparser==2.19
pyOpenSSL==19.1.0
PySocks==1.7.1
selenium==3.141.0
six==1.13.0
urllib3==1.25.7
win-inet-pton==1.1.0
wincertstore==0.2
